# core-lending-delinquency-collections-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-delinquency-collections
  description: Loan Delinquency & Collections Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-collections
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/collection-cases/{collectionCaseId}/promises/{promiseId}:
    get:
      tags:
        - CollectionPromise
      summary: Get a payment promise by ID
      operationId: getById_2
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: promiseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionPromiseDTO'
    put:
      tags:
        - CollectionPromise
      summary: Update a payment promise
      operationId: update_2
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: promiseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionPromiseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionPromiseDTO'
    delete:
      tags:
        - CollectionPromise
      summary: Delete a payment promise record
      operationId: delete_2
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: promiseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/collection-cases/{collectionCaseId}/promises:
    get:
      tags:
        - CollectionPromise
      summary: List or search promises for a collection case
      operationId: findAll_2
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.collectionCaseId
          in: query
          description: Exact filter for collection case id
          required: false
          schema:
            type: string
        - name: filters.promisedDate
          in: query
          description: Exact filter for promised date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[promisedDate].from
          in: query
          description: Filter promised date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[promisedDate].to
          in: query
          description: Filter promised date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.promisedAmount
          in: query
          description: Exact filter for promised amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[promisedAmount].from
          in: query
          description: Filter promised amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[promisedAmount].to
          in: query
          description: Filter promised amount to value
          required: false
          schema:
            type: string
        - name: filters.actualPaidDate
          in: query
          description: Exact filter for actual paid date
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[actualPaidDate].from
          in: query
          description: Filter actual paid date from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[actualPaidDate].to
          in: query
          description: Filter actual paid date to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.actualPaidAmount
          in: query
          description: Exact filter for actual paid amount
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[actualPaidAmount].from
          in: query
          description: Filter actual paid amount from value
          required: false
          schema:
            type: string
        - name: rangeFilters.ranges[actualPaidAmount].to
          in: query
          description: Filter actual paid amount to value
          required: false
          schema:
            type: string
        - name: filters.isKept
          in: query
          description: Exact filter for is kept
          required: false
          schema:
            type: boolean
        - name: filters.note
          in: query
          description: Exact filter for note
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseCollectionPromiseDTO'
    post:
      tags:
        - CollectionPromise
      summary: Create a new payment promise
      operationId: create_2
      parameters:
        - name: collectionCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionPromiseDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CollectionPromiseDTO'
...
components:
  schemas:
...
    CollectionPromiseDTO:
      required:
        - collectionCaseId
        - promisedAmount
        - promisedDate
      type: object
      properties:
        collectionPromiseId:
          type: string
          format: uuid
          readOnly: true
        collectionCaseId:
          type: string
          format: uuid
        promisedDate:
          type: string
          format: date-time
        promisedAmount:
          minimum: 0.01
          exclusiveMinimum: false
          type: number
        actualPaidDate:
          type: string
          format: date-time
        actualPaidAmount:
          minimum: 0
          exclusiveMinimum: false
          type: number
        isKept:
          type: boolean
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseCollectionPromiseDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/CollectionPromiseDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...